requirements.txt
Django==1.11.27
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.9.4
djangorestframework-stubs==1.0.0
dnspython==1.16.0
mongoengine==0.18.2
pymongo==3.9.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
